Healthcare_Assistant/requirements.txt
streamlit==1.28.0
openai==1.3.0
faiss-cpu==1.12.0
langchain==0.0.330
python-dotenv==1.0.0
PyPDF2==3.0.1
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3
langchain-community==0.0.15
